# chore(ebpf): optimize filldir64 program