@geyslan 2024-07-10
Awesome. ❤️

I have only one question: as `evaluate_scope_filters()` updates `p`, it could somehow jeopardise some expected behaviour?

Doubt:

Commit says:

`However, its usecase only requires the submission of events
from non 0 process inodes.`

Wouldn't it be the opposite? `... only ... from 0 process...`?