@NDStrahilevitz 2024-07-16
> I have only one question: as `evaluate_scope_filters()` updates `p`, it could somehow jeopardise some expected behaviour?

I don't think so, since this isn't a program_data struct which is shared later on.